refactor: Rely on PSR HTTP factory interface instead of concrete implementation.
JsonApiFactory had a concrete dependancy.
To improve usability and respect Dependancy Inversion Principle we chose to use interface instead.
It allows implementer to use the implementation he wants even if sensio/framework-extra-bundle combined
with symfony/psr-http-message-bridge hardly suggest to use nyholm/psr7.
That is why we think it is irrelevant to suggest an implementation because one of our requirements already dit that.